--- src/markdown-viewer/markdown-it.ts
@@ -1,36 +1,0 @@
-import MarkdownIt from "markdown-it";
-import { urlRewriterFactory, transformComponentLink,components } from "./web-components-transformer";
-import nestedHeadersRulePlugin from "./nested-headers-rule";
-import tabsRulePlugin from "./tabs-rule";
-import linkTransformRulerPlugin from "./link-transform-ruler";
-import { ListItem } from "./tabs-ruler";
-import WebComponentsDriverRulerPlugin from "./web-components-driver-ruler";
-
-const extensions = components.flatMap(c => c.extensions);
-
-const listItemIsTabPanel = (listItem: ListItem): boolean => {
-    return extensions.some(extension => listItem.getLink()?.getPath()?.endsWith(extension));
-};
-
-export default function mdFactory(src: string) {
-    const md = MarkdownIt();
-    md.use(nestedHeadersRulePlugin, {
-        isAriaExpanded: (level: number) => {
-            if (level > 2) {
-                return false;
-            }
-
-            return undefined;
-        }
-    });
-
-    md.use(tabsRulePlugin, {
-        listItemIsTabPanel
-    });
-
-    md.use(linkTransformRulerPlugin, { transformers: [ urlRewriterFactory(src), transformComponentLink ] });
-
-    md.use(WebComponentsDriverRulerPlugin);
-
-    return md;
-}
