+++ src/markdown-viewer/tabs-ruler/index.ts
@@ -1,11 +1,10 @@
 import MarkdownIt, { StateCore, Token } from "markdown-it";
-import { Link, LinkImpl } from "../link-transform-ruler";
 
-export interface tabsRulePluginOptions {
+export interface tabsRulerPluginOptions {
   listItemIsTabPanel?: (listItem: ListItem) => boolean;
 }
 
-export default function tabsRulePlugin(md: MarkdownIt, options?: tabsRulePluginOptions): void {
+export function tabsRulerPlugin(md: MarkdownIt, options?: tabsRulerPluginOptions): void {
   const listItemIsTabPanelDelegate = options?.listItemIsTabPanel || ((_listItem: ListItem) => false);
 
   md.core.ruler.push("tabs_ruler", (state: StateCore): void => {
@@ -72,8 +71,8 @@ function setTabsTokenProperty(i: number, tokens: Token[]) {
 
 
 export type ListItem = {
+  findToken(type: string): Token | undefined;
   tokens: Token[];
-  getLink(): Link | null;
 }
 
 class ListItemImpl implements ListItem {
@@ -83,41 +82,25 @@ class ListItemImpl implements ListItem {
     this.tokens = tokens;
   }
 
-  getLink(): Link | null {
-    const tokens = getTokenWithLink(this.tokens);
-    let inLink = false;
-    const linkTokens: Token[] = [];
+  findToken(type: string): Token | undefined {
+    return this._getToken(this.tokens, type);
+  }
+
+  private _getToken(tokens: Token[], type: string): Token | undefined {
+    const token = tokens.find(token => token.type === type);
+    if (token) {
+      return token;
+    }
 
     for (const token of tokens) {
-      if (token.type === "link_open") {
-        inLink = true;
-      }
-
-      if (inLink) {
-        linkTokens.push(token);
-      }
-
-      if (token.type === "link_close") {
-        inLink = false;
+      if (token.children) {
+        const child = this._getToken(token.children, type);
+        if (child) {
+          return child;
+        }
       }
     }
 
-    if (linkTokens.length === 0) {
-      return null;
-    }
-
-    return new LinkImpl(linkTokens);
-
-    function getTokenWithLink(tokens: Token[]): Token[] {
-      if (tokens.some(token => token.type === "link_open")) {
-        return tokens;
-      }
-
-      if (tokens.some(token => token.type === "inline" && token.children?.some(t => t.type === "link_open"))) {
-        return tokens.find(token => token.type === "inline")?.children!;
-      }
-
-      return [];
-    }
+    return undefined;
   }
 }
